Use uint8_t for #threads, #levels, and diffusion factor

+++ include/keymix.h
@@ -3,6 +3,6 @@
 
 #include "types.h"
 
-int keymix(byte *seed, byte *out, size_t seed_size, mixing_config *config, uint32_t nof_threads);
+int keymix(byte *seed, byte *out, size_t seed_size, mixing_config *config, uint8_t nof_threads);
 
 #endif
