+++ verify.c
@@ -34,9 +34,9 @@ byte *setup(size_t size, bool random) {
 }
 
 typedef struct {
-        void (*func)(thread_data *, uint32_t);
+        void (*func)(thread_data *, uint8_t);
         thread_data thr_data;
-        uint32_t level;
+        uint8_t level;
 } run_thr_t;
 
 void *_run_thr(void *a) {
@@ -45,14 +45,14 @@ void *_run_thr(void *a) {
         return NULL;
 }
 
-void emulate_shuffle_chunks(void (*func)(thread_data *, uint32_t), byte *out, byte *in, size_t size,
-                            uint32_t level, size_t fanout, uint32_t nof_threads) {
+void emulate_shuffle_chunks(void (*func)(thread_data *, uint8_t), byte *out, byte *in, size_t size,
+                            uint8_t level, uint8_t fanout, uint8_t nof_threads) {
         if (nof_threads > (size / SIZE_MACRO))
                 nof_threads = fanout;
 
-        uint32_t thread_levels = level - LOGBASE(nof_threads, fanout); // only accurate when the
-                                                                       // nof_threads is a power of
-                                                                       // fanout
+        uint8_t thread_levels = level - LOGBASE(nof_threads, fanout); // only accurate when the
+                                                                      // nof_threads is a power of
+                                                                      // fanout
 
         pthread_t threads[nof_threads];
         run_thr_t thread_args[nof_threads];
@@ -66,7 +66,7 @@ void emulate_shuffle_chunks(void (*func)(thread_data *, uint32_t), byte *out, by
         // function
         mixing_config mconf = {NULL, fanout};
 
-        for (uint32_t t = 0; t < nof_threads; t++) {
+        for (uint8_t t = 0; t < nof_threads; t++) {
                 run_thr_t *arg = thread_args + t;
 
                 thread_data thr_data = {
@@ -89,7 +89,7 @@ void emulate_shuffle_chunks(void (*func)(thread_data *, uint32_t), byte *out, by
                 pthread_create(&threads[t], NULL, _run_thr, arg);
         }
 
-        for (uint32_t t = 0; t < nof_threads; t++) {
+        for (uint8_t t = 0; t < nof_threads; t++) {
                 pthread_join(threads[t], NULL);
         }
 }
@@ -98,7 +98,7 @@ void emulate_shuffle_chunks(void (*func)(thread_data *, uint32_t), byte *out, by
 // fanout^level macro blocks.
 // Note, since shuffle and spread use two different mixing schemas these do
 // not produce the same results, hence we do not compare them.
-int verify_shuffles(size_t fanout, uint32_t level) {
+int verify_shuffles(size_t fanout, uint8_t level) {
         size_t size = (size_t)pow(fanout, level) * SIZE_MACRO;
 
         _log(LOG_INFO, "> Verifying swaps and shuffles up to level %zu (%.2f MiB)\n", level,
@@ -115,8 +115,8 @@ int verify_shuffles(size_t fanout, uint32_t level) {
         byte *out_spread3  = setup(size, false);
 
         int err = 0;
-        for (uint32_t l = 1; l <= level; l++) {
-                uint32_t nof_threads         = pow(fanout, fmin(l, 3));
+        for (uint8_t l = 1; l <= level; l++) {
+                uint8_t nof_threads          = pow(fanout, fmin(l, 3));
                 bool is_shuffle_chunks_level = (level - l < fmin(l, 3));
 
                 // Fill in buffer of the inplace operations
@@ -186,7 +186,7 @@ int verify_shuffles(size_t fanout, uint32_t level) {
 // fanout^level macro blocks with a varying number of threads.
 // Note, since shuffle and spread use two different mixing schemas these do
 // not produce the same results, hence we do not compare them.
-int verify_shuffles_with_varying_threads(size_t fanout, uint32_t level) {
+int verify_shuffles_with_varying_threads(size_t fanout, uint8_t level) {
         size_t size = (size_t)pow(fanout, level) * SIZE_MACRO;
 
         _log(LOG_INFO, "> Verifying that shuffles AT level %zu are thread-independent (%.2f MiB)\n",
@@ -281,7 +281,7 @@ int verify_shuffles_with_varying_threads(size_t fanout, uint32_t level) {
 
 // Verify the equivalence of the results when using different encryption
 // functions (i.e., AES-NI, OpenSSL, WolfSSL)
-int verify_encs(size_t fanout, uint32_t level) {
+int verify_encs(size_t fanout, uint8_t level) {
         size_t size = (size_t)pow(fanout, level) * SIZE_MACRO;
 
         _log(LOG_INFO, "> Verifying encryption for size %.2f MiB\n", MiB(size));
@@ -317,7 +317,7 @@ int verify_encs(size_t fanout, uint32_t level) {
 
 // Verify the equivalence of the results when using single-threaded and
 // multi-threaded encryption with a varying number of threads
-int verify_multithreaded_encs(size_t fanout, uint32_t level) {
+int verify_multithreaded_encs(size_t fanout, uint8_t level) {
         size_t size = (size_t)pow(fanout, level) * SIZE_MACRO;
 
         _log(LOG_INFO, "> Verifying keymix equivalence for size %.2f MiB\n", MiB(size));
@@ -360,7 +360,7 @@ int verify_multithreaded_encs(size_t fanout, uint32_t level) {
         return err;
 }
 
-int verify_keymix_t(size_t fanout, uint32_t level) {
+int verify_keymix_t(size_t fanout, uint8_t level) {
         size_t size = (size_t)pow(fanout, level) * SIZE_MACRO;
 
         _log(LOG_INFO, "> Verifying keymix-t equivalence for size %.2f MiB\n", MiB(size));
@@ -376,8 +376,8 @@ int verify_keymix_t(size_t fanout, uint32_t level) {
 
         mixing_config conf = {&aesni, fanout};
 
-        uint128_t iv              = rand() % (1 << sizeof(uint128_t));
-        uint32_t internal_threads = 1;
+        uint128_t iv             = rand() % (1 << sizeof(uint128_t));
+        uint8_t internal_threads = 1;
 
         // Note: Keymix T applies the IV, so we have to do that manually
         // to the input of Keymix
@@ -419,9 +419,9 @@ int main() {
 
         int err = 0;
 
-        for (size_t fanout = 2; fanout <= 4; fanout++) {
+        for (uint8_t fanout = 2; fanout <= 4; fanout++) {
                 _log(LOG_INFO, "Verifying with fanout %zu\n", fanout);
-                for (size_t l = MIN_LEVEL; l <= MAX_LEVEL; l++) {
+                for (uint8_t l = MIN_LEVEL; l <= MAX_LEVEL; l++) {
                         CHECKED(verify_shuffles(fanout, l));
                         CHECKED(verify_shuffles_with_varying_threads(fanout, l));
                         CHECKED(verify_encs(fanout, l));
